[feat]: serviço e domínio para cálculo de expressão matemática.

+++ src/VC.PorterGroup.TestePratico.CalculoDeExpressao.Dominio/Interface/IProcessadorMatematicoServico.cs
@@ -2,5 +2,5 @@
 
 public interface IProcessadorMatematicoServico
 {
-    decimal Calcular(string expressao);
+    double Calcular(string expressao);
 }
